Updated Result.java to aas4j
